3d/pulse: build and run the extern and static backend tests
The extern and static backends had never been extracted to C, because
share/everparse/tests/3d/Makefile never recursed into the copied
subdirectories. Doing so exposed several defects, all fixed here.

Wrapper codegen (Target.fst): emit a stream-shaped wrapper and error
handler under --pulse. The parsed size comes from EverParseStreamGetPosition,
whose spec already states that it is the number of consumed bytes, so no
position needs to be threaded through the validators. The extra argument
stays in the wrapper signature but is no longer forwarded.

Makefile-driven codegen (GenMakefile.fst, Batch.ml): under --pulse,
KaRaMeL EverParse.h is the real one and must not be deleted; it now
depends on every module .c so make does not consider it up to date before
krml runs; and EverParse.c gained rules and a place in the all-files lists
so EverParseErrorReasonOfResult resolves at link time.

field_ptr_after is allowed for static as well as extern, matching Low*,
where static is the extern development with different C linkage.

The two client streams are rewritten against the Pulse ABI.

+++ share/everparse/tests/3d/extern/src/main.c
@@ -17,9 +17,9 @@ int main(void) {
   if (test != NULL) {
     EVERPARSE_INPUT_STREAM_BASE testStream = EverParseCreate();
     if (testStream != NULL) {
-      EverParsePush(testStream, test, testSize);
-      EverParsePush(testStream, test, testSize);
-      EverParsePush(testStream, test, testSize);
+      EverParsePush(testStream, test, (size_t)testSize);
+      EverParsePush(testStream, test, (size_t)testSize);
+      EverParsePush(testStream, test, (size_t)testSize);
       if (TestCheckPoint(0, testStream)) {
         printf("Validation succeeded\n");
       }
